melona_diagnosis/requirements.txt
python==3.10.12
numpy==1.19.5
Pillow==7.2.0
matplotlib==3.4.3
torch==1.9.1
torchvision==0.10.1
opencv-python==4.5.3.56
scipy==1.7.1
scikit-image==0.18.3
pandas==1.3.2